@@ -2,8 +2,8 @@
 
     <PropertyGroup>
         <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-        <TargetFrameworks>net5.0;net6.0</TargetFrameworks>
         <LangVersion>preview</LangVersion>
+        <TargetFramework>net6.0</TargetFramework>
     </PropertyGroup>
 
 </Project>
